# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlimefishBackend API",
    "version": "1.0",
    "description": "[Discord](https://discord.gg/Q5CeSMFeBP) · [API key](https://slimefishbackend.dev/settings/api) · [OpenAPI spec](https://github.com/casesandberg/slimefish-backend/blob/main/apps/api/openapi.json) · [Trading bot](https://github.com/casesandberg/slimefishbackend-bot-trading) · [Mirror bot](https://github.com/casesandberg/slimefishbackend-bot-manifold-mirror)\n\n Include `x-api-key: {{key}}` in the header of your requests."
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "id": {
        "type": "string"
      },
...
    },
    "parameters": {
      "id": {
        "schema": {
          "allOf": [
            {
              "$ref": "#/components/schemas/id"
            },
            {
              "format": "cuid"
            }
          ]
        },
        "required": true,
        "name": "id",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/markets/[id]/resolve": {
      "post": {
        "summary": "Resolve a market",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "optionId": {
                    "type": "string"
                  },
                  "supportingLink": {
                    "type": "string"
                  }
                },
                "required": [
                  "optionId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "success"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}